Creacion del customHook para las validaciones
