+++ src/style/register.css
@@ -12,6 +12,7 @@
   color: azure;
   font-size: 54px;
   text-shadow: 1px 1px 2px rgba(60, 95, 155, 0.7);
+  user-select: none;
 }
 .register {
   display: flex;
@@ -21,18 +22,14 @@
 
 .register div img {
   max-width: 559px;
-  height: auto;
   border-top-right-radius: 20px;
   border-top-left-radius: 20px;
 }
 .register div form {
-  display: grid;
-  justify-items: center;
-  grid-template-columns: repeat(1, 0fr);
-  grid-gap: 16px;
-  position: relative;
-  bottom: 20px;
-  padding: 60px 30px 0 30px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 40px 30px 30px 30px;
   border-radius: 20px;
 }
 .register div form button {
@@ -41,9 +38,16 @@
   color: white;
   border: none;
   border-radius: 24px;
-  box-shadow: 0 4px 8px rgba(0, 0, 255, 0.652);
-  background-color: rgba(0, 0, 255, 0.652);
+  box-shadow: 0 4px 8px rgba(0, 0, 255, 0.6);
+  background-color: rgba(0, 0, 255, 0.6);
   cursor: pointer;
+}
+.register div form span {
+  font-size: large;
+  font-weight: 600;
+  color: red;
+  height: 10px;
+  margin-bottom: 24px;
 }
 .register div form div {
   display: grid;
